Return long instead of int when we are counting users by login.
Should be in 635dbc642eee9c1ae329e57bc03a27b0e92f02cd commit.

No functional changes.

+++ src/main/java/ru/mystamps/web/dao/UserDao.java
@@ -23,5 +23,5 @@ import ru.mystamps.web.entity.User;
 
 public interface UserDao extends CrudRepository<User, Integer> {
 	User findByLogin(String login);
-	int countByLogin(String login);
+	long countByLogin(String login);
 }
